More clean up work in the manifest loading logic
