+++ Sources/TuistLoader/Models+ManifestMappers/BuildConfiguration+ManifestMapper.swift
@@ -3,6 +3,10 @@ import ProjectDescription
 import TuistCore
 
 extension TuistCore.BuildConfiguration {
+    /// Maps a ProjectDescription.BuildConfiguration instance into a TuistCore.BuildConfiguration instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of build configuration model.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.CustomConfiguration) -> TuistCore.BuildConfiguration {
         let variant: TuistCore.BuildConfiguration.Variant
         switch manifest.variant {
