+++ Sources/TuistLoader/Models+ManifestMappers/CompatibleXcodeVersions+ManifestMapper.swift
@@ -5,6 +5,10 @@ import TuistCore
 import TuistSupport
 
 extension TuistCore.CompatibleXcodeVersions {
+    /// Maps a ProjectDescription.CompatibleXcodeVersions instance into a TuistCore.CompatibleXcodeVersions model.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of compatible Xcode versions.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.CompatibleXcodeVersions) -> TuistCore.CompatibleXcodeVersions {
         switch manifest {
         case .all:
