+++ Sources/TuistLoader/Models+ManifestMappers/CoreDataModel+ManifestMapper.swift
@@ -5,9 +5,11 @@ import TuistCore
 import TuistSupport
 
 extension TuistCore.CoreDataModel {
-    static func from(manifest: ProjectDescription.CoreDataModel,
-                     path _: AbsolutePath,
-                     generatorPaths: GeneratorPaths) throws -> TuistCore.CoreDataModel {
+    /// Maps a ProjectDescription.CoreDataModel instance into a TuistCore.CoreDataModel instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of Core Data model.
+    ///   - generatorPaths: Generator paths.
+    static func from(manifest: ProjectDescription.CoreDataModel, generatorPaths: GeneratorPaths) throws -> TuistCore.CoreDataModel {
         let modelPath = try generatorPaths.resolve(path: manifest.path)
         let versions = FileHandler.shared.glob(modelPath, glob: "*.xcdatamodel")
         let currentVersion = manifest.currentVersion
