+++ Sources/TuistLoader/Models+ManifestMappers/DefaultSettings+ManifestMapper.swift
@@ -5,6 +5,10 @@ import TuistCore
 import TuistSupport
 
 extension TuistCore.DefaultSettings {
+    /// Maps a ProjectDescription.DefaultSettings instance into a TuistCore.DefaultSettings model.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of default settings.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.DefaultSettings) -> TuistCore.DefaultSettings {
         switch manifest {
         case .recommended:
