+++ Sources/TuistLoader/Models+ManifestMappers/Dependency+ManifestMapper.swift
@@ -4,6 +4,10 @@ import ProjectDescription
 import TuistCore
 
 extension TuistCore.Dependency {
+    /// Maps a ProjectDescription.TargetDependency instance into a TuistCore.Dependency instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of the target dependency model.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.TargetDependency, generatorPaths: GeneratorPaths) throws -> TuistCore.Dependency {
         switch manifest {
         case let .target(name):
