+++ Sources/TuistLoader/Models+ManifestMappers/Headers+ManifestMapper.swift
@@ -5,30 +5,33 @@ import TuistCore
 import TuistSupport
 
 extension TuistCore.Headers {
-    static func from(manifest: ProjectDescription.Headers,
-                     path _: AbsolutePath,
-                     generatorPaths: GeneratorPaths) throws -> TuistCore.Headers {
+    /// Maps a ProjectDescription.Headers instance into a TuistCore.Headers model.
+    /// Glob patterns are resolved as part of the mapping process.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of Headers.
+    ///   - generatorPaths: Generator paths.
+    static func from(manifest: ProjectDescription.Headers, generatorPaths: GeneratorPaths) throws -> TuistCore.Headers {
+        let unfoldGlob: (AbsolutePath) -> [AbsolutePath] = { path in
+            FileHandler.shared.glob(AbsolutePath.root, glob: String(path.pathString.dropFirst())).filter {
+                if let fileExtension = $0.extension {
+                    return TuistCore.Headers.extensions.contains(".\(fileExtension)")
+                }
+                return false
+            }
+        }
+
         let `public` = try manifest.public?.globs.flatMap {
-            headerFiles(try generatorPaths.resolve(path: $0))
+            unfoldGlob(try generatorPaths.resolve(path: $0))
         } ?? []
 
         let `private` = try manifest.private?.globs.flatMap {
-            headerFiles(try generatorPaths.resolve(path: $0))
+            unfoldGlob(try generatorPaths.resolve(path: $0))
         } ?? []
 
         let project = try manifest.project?.globs.flatMap {
-            headerFiles(try generatorPaths.resolve(path: $0))
+            unfoldGlob(try generatorPaths.resolve(path: $0))
         } ?? []
 
         return Headers(public: `public`, private: `private`, project: project)
-    }
-
-    private static func headerFiles(_ path: AbsolutePath) -> [AbsolutePath] {
-        FileHandler.shared.glob(AbsolutePath("/"), glob: String(path.pathString.dropFirst())).filter {
-            if let `extension` = $0.extension, Headers.extensions.contains(".\(`extension`)") {
-                return true
-            }
-            return false
-        }
     }
 }
