+++ Sources/TuistLoader/Models+ManifestMappers/Platform+ManifestMapper.swift
@@ -3,6 +3,10 @@ import ProjectDescription
 import TuistCore
 
 extension TuistCore.Platform {
+    /// Maps a ProjectDescription.Platform instance into a TuistCore.Platform instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of platform model.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.Platform) throws -> TuistCore.Platform {
         switch manifest {
         case .macOS:
