+++ Sources/TuistLoader/Models+ManifestMappers/Product+ManifestMapper.swift
@@ -3,6 +3,10 @@ import ProjectDescription
 import TuistCore
 
 extension TuistCore.Product {
+    /// Maps a ProjectDescription.Product instance into a TuistCore.Product instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of product model.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.Product) -> TuistCore.Product {
         switch manifest {
         case .app:
