+++ Sources/TuistLoader/Models+ManifestMappers/Scheme+ManifestMapper.swift
@@ -4,42 +4,20 @@ import ProjectDescription
 import TuistCore
 
 extension TuistCore.Scheme {
-    static func from(manifest: ProjectDescription.Scheme, projectPath: AbsolutePath, generatorPaths: GeneratorPaths) throws -> TuistCore.Scheme {
+    /// Maps a ProjectDescription.Scheme instance into a TuistCore.Scheme instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of build action model.
+    ///   - generatorPaths: Generator paths.
+    static func from(manifest: ProjectDescription.Scheme, generatorPaths: GeneratorPaths) throws -> TuistCore.Scheme {
         let name = manifest.name
         let shared = manifest.shared
         let buildAction = try manifest.buildAction.map { try TuistCore.BuildAction.from(manifest: $0,
-                                                                                        projectPath: projectPath,
                                                                                         generatorPaths: generatorPaths) }
         let testAction = try manifest.testAction.map { try TuistCore.TestAction.from(manifest: $0,
-                                                                                     projectPath: projectPath,
                                                                                      generatorPaths: generatorPaths) }
         let runAction = try manifest.runAction.map { try TuistCore.RunAction.from(manifest: $0,
                                                                                   generatorPaths: generatorPaths) }
         let archiveAction = try manifest.archiveAction.map { try TuistCore.ArchiveAction.from(manifest: $0,
-                                                                                              projectPath: projectPath,
-                                                                                              generatorPaths: generatorPaths) }
-
-        return Scheme(name: name,
-                      shared: shared,
-                      buildAction: buildAction,
-                      testAction: testAction,
-                      runAction: runAction,
-                      archiveAction: archiveAction)
-    }
-
-    static func from(manifest: ProjectDescription.Scheme, workspacePath: AbsolutePath, generatorPaths: GeneratorPaths) throws -> TuistCore.Scheme {
-        let name = manifest.name
-        let shared = manifest.shared
-        let buildAction = try manifest.buildAction.map { try TuistCore.BuildAction.from(manifest: $0,
-                                                                                        projectPath: workspacePath,
-                                                                                        generatorPaths: generatorPaths) }
-        let testAction = try manifest.testAction.map { try TuistCore.TestAction.from(manifest: $0,
-                                                                                     projectPath: workspacePath,
-                                                                                     generatorPaths: generatorPaths) }
-        let runAction = try manifest.runAction.map { try TuistCore.RunAction.from(manifest: $0,
-                                                                                  generatorPaths: generatorPaths) }
-        let archiveAction = try manifest.archiveAction.map { try TuistCore.ArchiveAction.from(manifest: $0,
-                                                                                              projectPath: workspacePath,
                                                                                               generatorPaths: generatorPaths) }
 
         return Scheme(name: name,
