+++ Sources/TuistLoader/Models+ManifestMappers/SettingValue+ManifestMapper.swift
@@ -5,6 +5,10 @@ import TuistCore
 import TuistSupport
 
 extension TuistCore.SettingValue {
+    /// Maps a ProjectDescription.SettingValue instance into a TuistCore.SettingValue model.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of setting value.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.SettingValue) -> TuistCore.SettingValue {
         switch manifest {
         case let .string(value):
