+++ Sources/TuistLoader/Models+ManifestMappers/Settings+ManifestMapper.swift
@@ -6,13 +6,17 @@ import TuistCore
 extension TuistCore.Settings {
     typealias BuildConfigurationTuple = (TuistCore.BuildConfiguration, TuistCore.Configuration?)
 
-    static func from(manifest: ProjectDescription.Settings, path _: AbsolutePath, generatorPaths: GeneratorPaths) throws -> TuistCore.Settings {
+    /// Maps a ProjectDescription.Settings instance into a TuistCore.Settings instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of  the settings.
+    ///   - generatorPaths: Generator paths.
+    static func from(manifest: ProjectDescription.Settings, generatorPaths: GeneratorPaths) throws -> TuistCore.Settings {
         let base = manifest.base.mapValues(TuistCore.SettingValue.from)
         let configurations = try manifest.configurations
             .reduce([TuistCore.BuildConfiguration: TuistCore.Configuration?]()) { acc, val in
                 var result = acc
                 let variant = TuistCore.BuildConfiguration.from(manifest: val)
-                result[variant] = try TuistCore.Configuration.from(manifest: val.configuration, generatorPaths: generatorPaths)
+                result[variant] = try val.configuration.map { try TuistCore.Configuration.from(manifest: $0, generatorPaths: generatorPaths) }
                 return result
             }
         let defaultSettings = TuistCore.DefaultSettings.from(manifest: manifest.defaultSettings)
