+++ Sources/TuistLoader/Models+ManifestMappers/Target+ManifestMapper.swift
@@ -4,10 +4,12 @@ import ProjectDescription
 import TuistCore
 
 extension TuistCore.Target {
+    /// Maps a ProjectDescription.Target instance into a TuistCore.Target instance.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of  the target.
+    ///   - generatorPaths: Generator paths.
     // swiftlint:disable:next function_body_length
-    static func from(manifest: ProjectDescription.Target,
-                     path: AbsolutePath,
-                     generatorPaths: GeneratorPaths) throws -> TuistCore.Target {
+    static func from(manifest: ProjectDescription.Target, generatorPaths: GeneratorPaths) throws -> TuistCore.Target {
         let name = manifest.name
         let platform = try TuistCore.Platform.from(manifest: manifest.platform)
         let product = TuistCore.Product.from(manifest: manifest.product)
@@ -18,15 +20,14 @@ extension TuistCore.Target {
 
         let dependencies = try manifest.dependencies.map { try TuistCore.Dependency.from(manifest: $0, generatorPaths: generatorPaths) }
 
-        let infoPlist = try TuistCore.InfoPlist.from(manifest: manifest.infoPlist, path: path, generatorPaths: generatorPaths)
+        let infoPlist = try TuistCore.InfoPlist.from(manifest: manifest.infoPlist, generatorPaths: generatorPaths)
         let entitlements = try manifest.entitlements.map { try generatorPaths.resolve(path: $0) }
 
-        let settings = try manifest.settings.map { try TuistCore.Settings.from(manifest: $0, path: path, generatorPaths: generatorPaths) }
-        let sources = try TuistCore.Target.sources(projectPath: path,
-                                                   sources: manifest.sources?.globs.map { (glob: ProjectDescription.SourceFileGlob) in
-                                                       let globPath = try generatorPaths.resolve(path: glob.glob).pathString
-                                                       let excluding: [String] = try glob.excluding.compactMap { try generatorPaths.resolve(path: $0).pathString }
-                                                       return (glob: globPath, excluding: excluding, compilerFlags: glob.compilerFlags)
+        let settings = try manifest.settings.map { try TuistCore.Settings.from(manifest: $0, generatorPaths: generatorPaths) }
+        let sources = try TuistCore.Target.sources(sources: manifest.sources?.globs.map { (glob: ProjectDescription.SourceFileGlob) in
+            let globPath = try generatorPaths.resolve(path: glob.glob).pathString
+            let excluding: [String] = try glob.excluding.compactMap { try generatorPaths.resolve(path: $0).pathString }
+            return (glob: globPath, excluding: excluding, compilerFlags: glob.compilerFlags)
         } ?? [])
 
         let resourceFilter = { (path: AbsolutePath) -> Bool in
@@ -34,18 +35,20 @@ extension TuistCore.Target {
         }
         let resources = try (manifest.resources ?? []).flatMap {
             try TuistCore.FileElement.from(manifest: $0,
-                                           path: path,
                                            generatorPaths: generatorPaths,
                                            includeFiles: resourceFilter)
         }
 
-        let headers = try manifest.headers.map { try TuistCore.Headers.from(manifest: $0, path: path, generatorPaths: generatorPaths) }
+        let headers = try manifest.headers.map { try TuistCore.Headers.from(manifest: $0, generatorPaths: generatorPaths) }
 
         let coreDataModels = try manifest.coreDataModels.map {
-            try TuistCore.CoreDataModel.from(manifest: $0, path: path, generatorPaths: generatorPaths)
+            try TuistCore.CoreDataModel.from(manifest: $0, generatorPaths: generatorPaths)
         }
 
-        let actions = try manifest.actions.map { try TuistCore.TargetAction.from(manifest: $0, path: path, generatorPaths: generatorPaths) }
+        let actions = try manifest.actions.map {
+            try TuistCore.TargetAction.from(manifest: $0, generatorPaths: generatorPaths)
+        }
+
         let environment = manifest.environment
 
         return TuistCore.Target(name: name,
