+++ Sources/TuistLoader/Models+ManifestMappers/Workspace+ManifestMapper.swift
@@ -5,6 +5,10 @@ import TuistCore
 import TuistSupport
 
 extension TuistCore.Workspace {
+    /// Maps a ProjectDescription.Workspace instance into a TuistCore.Workspace model.
+    /// - Parameters:
+    ///   - manifest: Manifest representation of  workspace.
+    ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.Workspace,
                      path: AbsolutePath,
                      generatorPaths: GeneratorPaths,
@@ -19,6 +23,8 @@ extension TuistCore.Workspace {
                 }
 
             if projects.isEmpty {
+                // FIXME: This should be done in a linter.
+                // Before we can do that we have to change the linters to run with the TuistCore models and not the ProjectDescription ones.
                 Printer.shared.print(warning: "No projects found at: \(path.pathString)")
             }
 
@@ -26,12 +32,10 @@ extension TuistCore.Workspace {
         }
 
         let additionalFiles = try manifest.additionalFiles.flatMap {
-            try TuistCore.FileElement.from(manifest: $0,
-                                           path: path,
-                                           generatorPaths: generatorPaths)
+            try TuistCore.FileElement.from(manifest: $0, generatorPaths: generatorPaths)
         }
 
-        let schemes = try manifest.schemes.map { try TuistCore.Scheme.from(manifest: $0, workspacePath: path, generatorPaths: generatorPaths) }
+        let schemes = try manifest.schemes.map { try TuistCore.Scheme.from(manifest: $0, generatorPaths: generatorPaths) }
 
         return TuistCore.Workspace(path: path,
                                    name: manifest.name,
