+++ Tests/TuistCoreTests/Models/TargetTests.swift
@@ -59,11 +59,10 @@ final class TargetTests: TuistUnitTestCase {
         ])
 
         // When
-        let sources = try Target.sources(projectPath: temporaryPath,
-                                         sources: [
-                                             (glob: temporaryPath.appending(RelativePath("sources/**")).pathString, excluding: [], compilerFlags: nil),
-                                             (glob: temporaryPath.appending(RelativePath("sources/**")).pathString, excluding: [], compilerFlags: nil),
-                                         ])
+        let sources = try Target.sources(sources: [
+            (glob: temporaryPath.appending(RelativePath("sources/**")).pathString, excluding: [], compilerFlags: nil),
+            (glob: temporaryPath.appending(RelativePath("sources/**")).pathString, excluding: [], compilerFlags: nil),
+        ])
 
         // Then
         let relativeSources = sources.map { $0.path.relative(to: temporaryPath).pathString }
@@ -91,12 +90,11 @@ final class TargetTests: TuistUnitTestCase {
         ])
 
         // When
-        let sources = try Target.sources(projectPath: temporaryPath,
-                                         sources: [
-                                             (glob: temporaryPath.appending(RelativePath("sources/**")).pathString,
-                                              excluding: [temporaryPath.appending(RelativePath("sources/**/*Tests.swift")).pathString],
-                                              compilerFlags: nil),
-                                         ])
+        let sources = try Target.sources(sources: [
+            (glob: temporaryPath.appending(RelativePath("sources/**")).pathString,
+             excluding: [temporaryPath.appending(RelativePath("sources/**/*Tests.swift")).pathString],
+             compilerFlags: nil),
+        ])
 
         // Then
         let relativeSources = sources.map { $0.path.relative(to: temporaryPath).pathString }
@@ -132,12 +130,11 @@ final class TargetTests: TuistUnitTestCase {
         ]
 
         // When
-        let sources = try Target.sources(projectPath: temporaryPath,
-                                         sources: [
-                                             (glob: temporaryPath.appending(RelativePath("sources/**")).pathString,
-                                              excluding: excluding,
-                                              compilerFlags: nil),
-                                         ])
+        let sources = try Target.sources(sources: [
+            (glob: temporaryPath.appending(RelativePath("sources/**")).pathString,
+             excluding: excluding,
+             compilerFlags: nil),
+        ])
 
         // Then
         let relativeSources = sources.map { $0.path.relative(to: temporaryPath).pathString }
