+++ Tests/TuistLoaderTests/Models+ManifestMappers/Settings+ManifestMapperTests.swift
@@ -18,7 +18,7 @@ final class SettingsManifestMapperTests: TuistUnitTestCase {
         let manifest = ProjectDescription.Settings(base: ["base": .string("base")], debug: debug, release: release)
 
         // When
-        let model = try TuistCore.Settings.from(manifest: manifest, path: temporaryPath, generatorPaths: generatorPaths)
+        let model = try TuistCore.Settings.from(manifest: manifest, generatorPaths: generatorPaths)
 
         // Then
         XCTAssertSettingsMatchesManifest(settings: model, matches: manifest, at: temporaryPath, generatorPaths: generatorPaths)
