+++ Tests/TuistLoaderTests/Models+ManifestMappers/TargetAction+ManifestMapperTests.swift
@@ -1,0 +1,31 @@
+import Basic
+import Foundation
+import ProjectDescription
+import TuistCore
+import TuistSupport
+import XCTest
+
+@testable import TuistLoader
+@testable import TuistSupportTesting
+
+final class TargetActionManifestMapperTests: TuistUnitTestCase {
+    func test_from() throws {
+        // Given
+        let temporaryPath = try self.temporaryPath()
+        let generatorPaths = GeneratorPaths(manifestDirectory: temporaryPath)
+        let manifest = ProjectDescription.TargetAction.test(name: "MyScript",
+                                                            tool: "my_tool",
+                                                            path: "my/path",
+                                                            order: .pre,
+                                                            arguments: ["arg1", "arg2"])
+        // When
+        let model = try TuistCore.TargetAction.from(manifest: manifest, generatorPaths: generatorPaths)
+
+        // Then
+        XCTAssertEqual(model.name, "MyScript")
+        XCTAssertEqual(model.tool, "my_tool")
+        XCTAssertEqual(model.path, temporaryPath.appending(RelativePath("my/path")))
+        XCTAssertEqual(model.order, .pre)
+        XCTAssertEqual(model.arguments, ["arg1", "arg2"])
+    }
+}
